+++ src/main/java/edu/kis/vh/nursery/DefaultCountingOutRyhmer.java
@@ -1,7 +1,7 @@
 package edu.kis.vh.nursery;
 
-import edu.kis.vh.nursery.list.IntLinkedList;
-import edu.kis.vh.nursery.list.Stack;
+import edu.kis.vh.nursery.stack.IntLinkedList;
+import edu.kis.vh.nursery.stack.Stack;
 
 public class DefaultCountingOutRyhmer {
