+++ src/main/java/edu/kis/vh/nursery/stack/IntArrayStack.java
@@ -1,6 +1,4 @@
-package edu.kis.vh.nursery;
-
-import edu.kis.vh.nursery.list.Stack;
+package edu.kis.vh.nursery.stack;
 
 public class IntArrayStack implements Stack {
     private static final int NUMBERS_CAPACITY = 12;
